api/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "langchain-ocr"
version = "0.1.0"
description = "Langchain OCR API"
authors = ["Andreas Klos"]
packages = [{ include = "langchain_ocr", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.30.0"
fastapi = "^0.111.0"
requests-oauthlib = "^2.0.0"
deprecated = "^1.2.14"
pyyaml = "^6.0.2"
pillow = "^11.0.0"
pytest-asyncio = "^0.25.0"
pycountry = "^24.6.1"
pdf2image = "^1.17.0"
inject = "^5.2.1"
langchain-ocr-lib = {path = "../langchain_ocr_lib"}


[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.1"
pytest = "^8.2.1"
coverage = "^7.5.4"
flake8 = "^7.1.0"
flake8-black = "^0.3.6"
flake8-pyproject = "^1.2.3"
flake8-quotes = "^3.4.0"
flake8-return = "^1.2.0"
flake8-annotations-complexity = "^0.0.8"
flake8-bandit = "^4.1.1"
flake8-bugbear = "^24.8.19"
flake8-builtins = "^2.5.0"
flake8-comprehensions = "^3.15.0"
flake8-eradicate = "^1.5.0"
flake8-expression-complexity = "^0.0.11"
flake8-pytest-style = "^2.0.0"
pep8-naming = "^0.14.1"
flake8-eol = "^0.0.8"
flake8-exceptions = "^0.0.1a0"
flake8-simplify = "^0.21.0"
flake8-wot = "^0.2.0"
flake8-function-order = "^0.0.5"
flake8-tidy-imports = "^4.10.0"
black = "^23.9.1"
# flake8-logging-format = "^2024.24.12"
flake8-docstrings = "^1.7.0"


[tool.flake8]
exclude= [".eggs", "src/rag_core_api/models", ".git", ".hg", ".mypy_cache", ".tox", ".venv", ".devcontainer", "venv", "_build", "buck-out", "build", "dist", "**/__init__.py"]
statistics = true
show-source = false
max-complexity = 8
max-annotations-complexity = 3
docstring-convention = 'numpy'
max-line-length = 120
ignore = ["E203", "W503", "E704"]
inline-quotes = '"'
docstring-quotes = '"""'
multiline-quotes = '"""'
dictionaries = ["en_US", "python", "technical", "pandas"]
ban-relative-imports = true
per-file-ignores = """

"""

[tool.black]
line-length = 120
exclude = """
/(
     .eggs
    | .git
    | .hg
    | .mypy_cache
    | .nox
    | .pants.d
    | .tox
    | .venv
    | _build
    | buck-out
    | build
    | dist
    | node_modules
    | venv
)/
"""

[tool.isort]
profile = "black"
skip = ['.eggs', '.git', '.hg', '.mypy_cache', '.nox', '.pants.d', '.tox', '.venv', '_build', 'buck-out', 'build', 'dist', 'node_modules', 'venv']
skip_gitignore = true
known_local_folder = ["langchain_ocr"]

[tool.pylint]
max-line-length = 120

[tool.pytest.ini_options]
log_cli = 1
log_cli_level = "DEBUG"
testpaths = "tests"
